refactor: V2サフィックスを削除してEditableTableを正式版に置き換え
## 変更内容

### ファイル削除
- 旧App.tsx、EditableTable.tsx、EditableTable.module.scssを削除
- 旧storiesディレクトリ（app/、EditableTable.stories.tsx）を削除

### ファイルリネーム
- AppV2.tsx → App.tsx
- EditableTableV2/ → EditableTable/
- EditableTableV2.tsx → EditableTable/index.tsx
- EditableTableV2.module.scss → EditableTable/index.module.scss
- useCellSelectionV2.ts → useCellSelection.ts
- useContextMenusV2.ts → useContextMenus.ts
- useTableSearchV2.ts → useTableSearch.ts
- stories/appV2/ → stories/app/
- App*.stories.tsx のリネーム（AppV2 → App）
- test ファイルのリネーム

### コード更新
- すべてのimport文でV2サフィックスを削除
- EditableTableV2Props → EditableTableProps
- コンポーネント名、関数名からV2を削除
- .markuplintrcのoverridesパスを更新

### 構造変更
- EditableTableをフォルダ構造に変更
- サブコンポーネント（EditableCell、HeaderCellなど）をEditableTable/配下に配置
- index.tsxとindex.module.scssを使用

## 備考
- markuplintのHeaderCellエラーは既知の問題（React componentsの認識）
- 型チェックは成功
- ESLintは成功
