--- webview-ui/src/components/EditableTable.module.scss
@@ -1,18 +1,0 @@
-.contextMenu {
-  position: absolute;
-  z-index: 2;
-}
-
-.dataGrid {
-  width: 100%;
-  height: calc(100% - 32px - 32px - 32px);
-  resize: "both";
-  > div[role="row"] > div[role="gridcell"] {
-    white-space: pre-wrap;
-  }
-}
-
-.divider {
-  width: 100%;
-  margin: 16px 0;
-}
