+++ webview-ui/src/components/EditableTable/DraggableHeaderCell.module.scss
