+++ webview-ui/src/components/EditableTable/DraggableHeaderCell.tsx
