+++ webview-ui/src/components/EditableTable/FilterInput.module.scss
