+++ webview-ui/stories/app/App.HeaderActions.stories.tsx
@@ -28,20 +28,39 @@ type Story = StoryObj<typeof meta>;
 
 async function triggerHeaderContextMenu(canvasElement: HTMLElement, targetValue: string) {
   const canvas = within(canvasElement);
-  // セルを右クリックしてコンテキストメニューを表示
-  const ageHeaderCell = await canvas.findByRole("columnheader", { name: "Age" });
+  // ヘッダーセルを右クリックしてコンテキストメニューを表示
+  const ageHeaderCell = await canvas.findByRole("columnheader", { name: /Age/ });
   await userEvent.click(ageHeaderCell);
   await userEvent.pointer({
     keys: "[MouseRight]",
     target: ageHeaderCell,
   });
-  const contextMenu = document.body.getElementsByTagName("vscode-context-menu");
-  await expect(contextMenu).toHaveLength(1);
-  const menuItem = contextMenu[0].shadowRoot?.querySelectorAll(
-    `vscode-context-menu-item[value='${targetValue}']`
+
+  // VscodeContextMenuが表示されるまで待つ
+  await waitFor(
+    async () => {
+      const contextMenu = document.body.querySelector("vscode-context-menu");
+      await expect(contextMenu).toBeInTheDocument();
+      return true;
+    },
+    { timeout: 2000 }
   );
-  const atag = menuItem?.[0].shadowRoot?.querySelector("a");
-  await userEvent.click(atag as HTMLElement);
+
+  // メニューアイテムをクリック（Shadow DOMを通じてアクセス）
+  const contextMenu = document.body.querySelector("vscode-context-menu");
+  const menuItem = contextMenu?.shadowRoot?.querySelector(
+    `vscode-context-menu-item[value="${targetValue}"]`
+  );
+  if (menuItem) {
+    const link = menuItem.shadowRoot?.querySelector("a");
+    if (link) {
+      await userEvent.click(link as HTMLElement);
+    } else {
+      throw new Error(`Link not found in menu item "${targetValue}"`);
+    }
+  } else {
+    throw new Error(`Menu item with value "${targetValue}" not found`);
+  }
 }
 
 export const AddHeaderForLeft: Story = {
@@ -56,13 +75,13 @@ export const AddHeaderForLeft: Story = {
     await waitFor(
       async () => {
         // 左側に新しいヘッダーが追加されていることを確認
-        const newHeader = newCanvas.getByRole("columnheader", { name: "new column" });
+        const newHeader = newCanvas.getByRole("columnheader", { name: /new column/ });
         await expect(newHeader).toBeInTheDocument();
 
         const headers = newCanvas.getAllByRole("columnheader");
         await expect(headers).toHaveLength(COL_MAX_WITH_HEADER + 1); // 新しいヘッダーが追加されていることを確認
 
-        await expect(headers[2].innerHTML).toContain("new column"); // 新しいヘッダーの内容を確認
+        await expect(headers[2].textContent).toContain("new column"); // 新しいヘッダーの内容を確認
         return true;
       },
       { timeout: 2000 }
@@ -81,13 +100,13 @@ export const AddHeaderForRight: Story = {
     const newCanvas = within(document.body);
     await waitFor(
       async () => {
-        // 左側に新しいヘッダーが追加されていることを確認
-        const newHeader = newCanvas.getByRole("columnheader", { name: "new column" });
+        // 右側に新しいヘッダーが追加されていることを確認
+        const newHeader = newCanvas.getByRole("columnheader", { name: /new column/ });
         await expect(newHeader).toBeInTheDocument();
 
         const headers = newCanvas.getAllByRole("columnheader");
         await expect(headers).toHaveLength(COL_MAX_WITH_HEADER + 1); // 新しいヘッダーが追加されていることを確認
-        await expect(headers[3].innerHTML).toContain("new column"); // 新しいヘッダーの内容を確認
+        await expect(headers[3].textContent).toContain("new column"); // 新しいヘッダーの内容を確認
         return true;
       },
       { timeout: 2000 }
@@ -108,7 +127,7 @@ export const DeleteHeader: Story = {
       async () => {
         const headers = newCanvas.getAllByRole("columnheader");
         await expect(headers).toHaveLength(COL_MAX_WITH_HEADER - 1);
-        await expect(headers[2].innerHTML).toContain("City");
+        await expect(headers[2].textContent).toContain("City");
         return true;
       },
       { timeout: 2000 }
