+++ webview-ui/stories/app/App.HeaderEditing.stories.tsx
@@ -3,6 +3,21 @@ import { expect, userEvent, within, waitFor } from "storybook/test";
 import App from "../../src/App";
 import { setInitData, waitReadyForGrid } from "./utils";
 
+/**
+ * EditableTableのヘッダーセル編集機能をテストするストーリー。
+ *
+ * 実装済み機能:
+ * - ダブルクリックでヘッダー編集モードに入る
+ * - F2キーでヘッダー編集モードに入る
+ * - Backspace/Deleteキーでヘッダーをクリア
+ * - 文字入力で編集モードに入る
+ * - Enter/Escapeキーで編集を確定/キャンセル
+ *
+ * 実装方法:
+ * - useHeaderEditingカスタムフックによる状態管理
+ * - EditableTableコンポーネントに統合
+ */
+
 const meta: Meta<typeof App> = {
   title: "App/HeaderEditing",
   component: App,
@@ -35,7 +50,7 @@ export const HeaderEditingFunctionality: Story = {
     await waitReadyForGrid(canvasElement);
 
     // ヘッダーセルをダブルクリックして編集モードに入る
-    const nameHeader = await canvas.findByRole("columnheader", { name: "Name" });
+    const nameHeader = await canvas.findByRole("button", { name: /Name/ });
     await userEvent.dblClick(nameHeader);
 
     // 編集可能なtextarea要素が表示されることを確認
@@ -60,7 +75,7 @@ export const HeaderEditingFunctionality_Backspace: Story = {
     await waitReadyForGrid(canvasElement);
 
     // ヘッダーセルをクリックして選択状態にする
-    const nameHeader = await canvas.findByRole("columnheader", { name: "Name" });
+    const nameHeader = await canvas.findByRole("button", { name: /Name/ });
     await userEvent.click(nameHeader);
 
     // ヘッダーセルが選択された状態でBackspaceキーを押す
@@ -89,9 +104,9 @@ export const HeaderEditingFunctionality_Delete: Story = {
     setInitData();
     await waitReadyForGrid(canvasElement);
 
-    // ヘッダーセルをクリックして選択状態にする
-    const nameHeader = await canvas.findByRole("columnheader", { name: "Name" });
-    await userEvent.click(nameHeader);
+    // ヘッダーセルのボタンをクリックしてフォーカスを当てる
+    const nameHeaderButton = await canvas.findByRole("button", { name: /Name/ });
+    await userEvent.click(nameHeaderButton);
 
     // ヘッダーセルが選択された状態でDeleteキーを押す
     await userEvent.keyboard("{Delete}");
@@ -105,14 +120,9 @@ export const HeaderEditingFunctionality_Delete: Story = {
     await expect(textarea).not.toBeInTheDocument();
 
     // ヘッダーセルの内容が空になっていることを確認
-    await waitFor(
-      async () => {
-        const updatedHeaders = canvas.queryAllByRole("columnheader", { name: "" });
-        await expect(updatedHeaders).toHaveLength(2); // インデックス行＋削除した列
-        return true;
-      },
-      { timeout: 1000 }
-    );
+    // "Name"というテキストを含むヘッダーが存在しないことを確認
+    const nameHeaderAfterDelete = canvas.queryByRole("button", { name: /Name/ });
+    await expect(nameHeaderAfterDelete).not.toBeInTheDocument();
   },
 };
 
@@ -123,7 +133,7 @@ async function insertHeaderTextTest(canvasElement: HTMLElement, key: string) {
   await waitReadyForGrid(canvasElement);
 
   // ヘッダーセルをクリックして選択状態にする
-  const nameHeader = await canvas.findByRole("columnheader", { name: "Name" });
+  const nameHeader = await canvas.findByRole("button", { name: /Name/ });
   await userEvent.click(nameHeader);
 
   // ヘッダーセルが選択された状態で文字を入力する
@@ -159,7 +169,7 @@ export const HeaderEditingFunctionality_F2Key: Story = {
     await waitReadyForGrid(canvasElement);
 
     // ヘッダーセルをクリックして選択状態にする
-    const nameHeader = await canvas.findByRole("columnheader", { name: "Name" });
+    const nameHeader = await canvas.findByRole("button", { name: /Name/ });
     await userEvent.click(nameHeader);
 
     // F2キーで編集モードに入る
@@ -189,7 +199,7 @@ export const HeaderEditingFunctionality_SpecialKeyInput: Story = {
     await waitReadyForGrid(canvasElement);
 
     // ヘッダーセルをクリックして選択状態にする
-    const nameHeader = await canvas.findByRole("columnheader", { name: "Name" });
+    const nameHeader = await canvas.findByRole("button", { name: /Name/ });
     await userEvent.click(nameHeader);
 
     // 以下のキーは編集モードに入らない（特殊動作をする）
@@ -220,7 +230,7 @@ export const HeaderEditingFunctionality_SaveChanges: Story = {
     await waitReadyForGrid(canvasElement);
 
     // ヘッダーセルをダブルクリックして編集モードに入る
-    const nameHeader = await canvas.findByRole("columnheader", { name: "Name" });
+    const nameHeader = await canvas.findByRole("button", { name: /Name/ });
     await userEvent.dblClick(nameHeader);
 
     // 編集可能なtextarea要素が表示されることを確認
@@ -244,8 +254,9 @@ export const HeaderEditingFunctionality_SaveChanges: Story = {
     // 編集モードが終了し、新しい内容が反映されることを確認
     await waitFor(
       async () => {
-        const updatedHeader = canvas.queryByRole("columnheader", { name: "新しい名前" });
+        const updatedHeader = canvas.queryByRole("columnheader", { name: /新しい名前/ });
         await expect(updatedHeader).toBeInTheDocument();
+        await expect(updatedHeader?.textContent).toContain("新しい名前");
         return true;
       },
       { timeout: 2000 }
