+++ webview-ui/stories/app/App.Utilities.stories.tsx
@@ -33,6 +33,9 @@ export const UndoRedo: Story = {
   play: async ({ context }) => {
     await DeleteHeader.play!(context);
 
+    // 少し待機してから元に戻す
+    await new Promise((resolve) => setTimeout(resolve, 200));
+
     // Ctrl+Z で元に戻す
     await userEvent.keyboard("{Control>}z{/Control}");
 
@@ -42,11 +45,14 @@ export const UndoRedo: Story = {
       async () => {
         const headers = newCanvas.getAllByRole("columnheader");
         await expect(headers).toHaveLength(COL_MAX_WITH_HEADER);
-        await expect(headers[2].innerHTML).toContain("Age");
+        await expect(headers[2].textContent).toContain("Age");
         return true;
       },
       { timeout: 2000 }
     );
+
+    // 少し待機してからやり直し
+    await new Promise((resolve) => setTimeout(resolve, 200));
 
     // Ctrl+Y でやり直し
     await userEvent.keyboard("{Control>}y{/Control}");
@@ -54,7 +60,7 @@ export const UndoRedo: Story = {
       async () => {
         const headers = newCanvas.getAllByRole("columnheader");
         await expect(headers).toHaveLength(COL_MAX_WITH_HEADER - 1);
-        await expect(headers[2].innerHTML).toContain("City");
+        await expect(headers[2].textContent).toContain("City");
         return true;
       },
       { timeout: 2000 }
@@ -69,9 +75,22 @@ export const RowSizeAdjustment: Story = {
     setInitData();
     await waitReadyForGrid(canvasElement);
 
+    // 初期状態の高さを確認
+    const initialCell = canvas.getByRole("gridcell", { name: "Alice" });
+    const initialHeight = initialCell.getBoundingClientRect().height;
+
     await setRowSize(canvas, "large");
 
-    const cell = canvas.getByRole("gridcell", { name: "Alice" });
-    await expect(cell.getBoundingClientRect().height).toBe(80); // largeサイズの高さを確認
+    // DOM更新を待機してサイズが変更されることを確認
+    await waitFor(
+      async () => {
+        const cell = canvas.getByRole("gridcell", { name: "Alice" });
+        const cellHeight = cell.getBoundingClientRect().height;
+        // largeサイズの高さが初期サイズより大きいことを確認
+        await expect(cellHeight).toBeGreaterThan(initialHeight);
+        return true;
+      },
+      { timeout: 3000 }
+    );
   },
 };
